move debugger to its own file
